Attack Picto: Get it working as intended.
Fortunately, the Edge of Eternities station mechanic has given us the ideal template to base our behavior from as we also want to enable a certain set of buffs and abilities based on whether a certain type of counter is on the equipment.

The important thing to note for other Pictos at this tier to be implemented: The Affected$ must be Card.Self+any_other_qualifiers, otherwise the continuous effect or trigger being installed by this equipment will be installed *for each* instance of the equipment you control, which will cause things the double-stack if you have a cloned instance.